+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 - Initial release.
 - Added bunch of decorations.
 - Added bunch of worldgen features.
+
+### 1.0.0-beta.4
+
+- Improved dynamic language generation.
+- Improved Biomes You'll Go compatibility.